Main stack implementation preparation
- Main stack implementation preparation
- Code refactoring
- Filtering meal from API preparation
