+++ Desktop/cookbook/src/data/meal/mod.rs
@@ -1,8 +1,9 @@
 pub mod area;
 pub mod categories;
+pub mod filtered;
 
-use crate::models::meal::api::{AllMealsAPI, MealAPI};
-use crate::models::meal::recipe::MealRecipe;
+use crate::dto::meal::{filtered_list::MealFilteredList, recipe::MealRecipe};
+use crate::models::meal::{AllMealsAPI, MealAPI};
 
 use reqwest::blocking::get;
