+++ Desktop/cookbook/src/dto/drink/mod.rs
@@ -1,0 +1,5 @@
+pub mod alcoholic_list;
+pub mod categories_list;
+pub mod glass_list;
+pub mod ingredient_list;
+pub mod recipe;
