+++ Desktop/cookbook/src/dto/meal/mod.rs
@@ -1,0 +1,5 @@
+pub mod area_list;
+pub mod categories_list;
+pub mod filtered_list;
+pub mod ingredient_list;
+pub mod recipe;
