+++ Desktop/cookbook/src/dto/mod.rs
@@ -1,0 +1,2 @@
+pub mod drink;
+pub mod meal;
