+++ Desktop/cookbook/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod data;
+pub mod dto;
 pub mod models;
