+++ Desktop/cookbook/src/models/drink/alcoholic.rs
