+++ Desktop/cookbook/src/models/drink/categories.rs
