+++ Desktop/cookbook/src/models/drink/ingredient.rs
