+++ Desktop/cookbook/src/models/meal/area.rs
