+++ Desktop/cookbook/src/models/meal/categories.rs
