+++ Desktop/cookbook/src/models/meal/ingredient.rs
