+++ Desktop/gui/src/gui_data/connections/buttons/random_drink.rs
@@ -1,5 +1,6 @@
 use crate::gui_data::{connections::buttons::common, GuiData};
-use cookbook::{data::drink::get_random_drink_recipe, models::drink::recipe::DrinkRecipe};
+use cookbook::data::drink::get_random_drink_recipe;
+use cookbook::dto::drink::recipe::DrinkRecipe;
 
 use gtk::prelude::*;
