+++ Desktop/gui/src/gui_data/connections/mod.rs
@@ -1,1 +1,2 @@
 pub mod buttons;
+pub mod tree_selections;
