+++ Desktop/gui/src/gui_data/main_window/buttons.rs
@@ -4,6 +4,7 @@ use gtk::prelude::*;
 pub struct MainWindowButtons {
     pub random_meal_recipe_button: gtk::Button,
     pub random_drink_recipe_button: gtk::Button,
+    pub previous_stack_button: gtk::Button,
 }
 
 impl MainWindowButtons {
@@ -14,10 +15,14 @@ impl MainWindowButtons {
         let random_drink_recipe_button: gtk::Button = builder
             .get_object("random_drink_recipe_button")
             .expect("\"random_drink_recipe_button\" ID in \"Main_Window.glade\" should exist.");
+        let previous_stack_button: gtk::Button = builder
+            .get_object("previous_stack_button")
+            .expect("\"previous_stack_button\" ID in \"Main_Window.glade\" should exist.");
 
         Self {
             random_meal_recipe_button,
             random_drink_recipe_button,
+            previous_stack_button,
         }
     }
 }
@@ -34,10 +39,14 @@ fn create_from_builder_test() {
     let random_drink_recipe_button: gtk::Button = builder
         .get_object("random_drink_recipe_button")
         .expect("\"random_drink_recipe_button\" ID in \"Main_Window.glade\" should exist.");
+    let previous_stack_button: gtk::Button = builder
+        .get_object("previous_stack_button")
+        .expect("\"previous_stack_button\" ID in \"Main_Window.glade\" should exist.");
 
     let should_be = MainWindowButtons {
         random_meal_recipe_button,
         random_drink_recipe_button,
+        previous_stack_button,
     };
 
     let tested = MainWindowButtons::create_from_builder(&builder);
